Updates Python interface for OpenMP/em2d version
- 🎉 Weibel test can now be run entirely from Python
- Current compilation is set for OpenMP + ARM/Neon (Apple Silicon)
- **warning** Class names may change in the future

+++ openmp/em2d/cathode.h
@@ -43,7 +43,7 @@ public:
         edge::pos wall, float ufl );
 
     void initialize( float2 const box, uint2 const ntiles, uint2 const nx,
-        float const dt, int const id_ ) override;
+        double const dt, int const id_ ) override;
     
     ~Cathode();
